# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Luna DevOps API
  version: 0.1.0
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/auth/providers/{providerId}:
    put:
      operationId: updateAuthProvider
      tags: [Auth]
      summary: Update auth provider
      parameters:
        - $ref: "#/components/parameters/ProviderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthProviderInput"
      responses:
        "200":
          description: Updated auth provider.
...
  /api/v1/git/providers/{providerId}:
    put:
      operationId: updateGitProvider
      tags: [Git]
      summary: Update Git provider
      parameters:
        - $ref: "#/components/parameters/ProviderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GitProviderInput"
      responses:
        "200":
          description: Updated Git provider.
    delete:
      operationId: deleteGitProvider
      tags: [Git]
      summary: Delete Git provider
      parameters:
        - $ref: "#/components/parameters/ProviderId"
      responses:
        "204":
          description: Deleted Git provider.
...
  /api/v1/configs/ai/provider/test:
    post:
      operationId: testAIProviderConnection
      tags: [AI Assistant]
      summary: Ask Agent to perform a minimal read-only Provider connection test
      security: [{ SessionCookie: [] }]
      responses:
        "200": { description: Stable Provider connection result, content: { application/json: { schema: { $ref: "#/components/schemas/AIObject" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "503": { description: Agent or Provider is unavailable, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: lyd_session
      description: HttpOnly browser session cookie. Sensitive interactive operations do not accept personal access tokens as a substitute.
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: Luna OAuth access token or personal access token
      description: Luna OAuth access token or personal access token. Individual operations may require OAuth and explicitly reject personal access tokens.
    AIAgentServiceToken:
      type: http
      scheme: bearer
      bearerFormat: independent Agent callback service credential
    AIDelegationToken:
      type: http
      scheme: bearer
      bearerFormat: 60-second operation-bound AI delegation
  parameters:
...
    ProviderId:
      name: providerId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Unauthorized:
      description: A valid interactive browser session is required.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    Forbidden:
      description: The authenticated principal is not allowed to perform this operation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    AIObject:
      type: object
      additionalProperties: true
      description: Versioned AI Agent response projected through the Luna API BFF.
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: Stable machine-readable error code.
        error:
          type: string
          description: Safe user-facing error message.
        detail:
          type: string
          description: Development-only diagnostic detail.
        purpose:
          type: string
          description: Step-up purpose returned with `mfa_required`.
        details:
          type: object
          additionalProperties: true
          description: Optional machine-readable error context. Scope failures include `requiredScope`.
      required: [code, error]
...
    AuthProviderInput:
      type: object
      properties:
        type: { type: string, enum: [oidc] }
        name: { type: string }
        enabled: { type: boolean }
        issuerUrl: { type: string }
        clientId: { type: string }
        clientSecret: { type: string }
        scopes: { type: string }
        groupClaim: { type: string }
        emailClaim: { type: string }
        usernameClaim: { type: string }
        isDefault: { type: boolean }
      required: [name, issuerUrl, clientId]
...
    GitProviderInput:
      type: object
      properties:
        type: { type: string, enum: [github, gitea, gitlab] }
        name: { type: string }
        baseUrl: { type: string }
        scope: { type: string, enum: [global, project, user] }
        ownerRef: { type: string }
        authType: { type: string, enum: [oauth, github-app, pat] }
        clientId: { type: string }
        clientSecret: { type: string, writeOnly: true }
        enabled: { type: boolean }
      required: [name]
...